// app-admin/src/app/company/models/company.ts
export class Company {
    id: number;
    alias: string;
    name_en: string;
    name_he: string;
    name_es: string;
    company_site_url: string;
    logo: number;
    image_company: number;
    image_product_list: number;
    color_1: string;
    color_2: string;
    color_3: string;
    color_4: string;
    color_5: string;
    status: number;
    created_at: number;
    created_by: number;
    updated_at: number;
    updated_by: number;
    deleted_at: number;
    deleted_by: number;
    activeRecipeCount?: string;

}